@@ -1,27 +1,32 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { Routes, RouterModule } from '@angular/router';
 
-import { RouterModule, Routes } from '@angular/router';
+import { AuthGuard } from './guard/auth-guard';
 
 import { LoginComponent } from './login/login.component';
-import { ChatComponent } from './chat/chat.component';
 import { RegisterComponent } from './register/register.component';
 
 const routes: Routes = [
-  { path: 'login', component: LoginComponent },
-  { path: 'register', component: RegisterComponent },
-  { path: '', component: ChatComponent },
-  { path: '**', redirectTo: '' }
-  // { path: '**', component: PageNotFoundComponent }
+  {
+    path: 'login',
+    component: LoginComponent,
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'register',
+    component: RegisterComponent,
+    canActivate: [AuthGuard]
+  },
+  {
+    path: '',
+    loadChildren: './chat/chat.module#ChatModule',
+    canActivate: [AuthGuard]
+},
 ];
 
 @NgModule({
-  imports: [
-    RouterModule.forRoot(
-      routes,
-      // { enableTracing: true } // <-- debugging purposes only
-    )
-  ],
-  exports: [RouterModule]
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule],
+  providers: [AuthGuard]
 })
 export class AppRoutingModule { }
